% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{bbox_coords_mat}
\alias{bbox_coords_mat}
\title{Bounding box as a polygon}
\usage{
bbox_coords_mat(polygon)
}
\arguments{
\item{polygon}{}
}
\value{
Coordinate matrix of a bounding box in the form of a polygon
}
\description{
Bounding box as a polygon
}
