% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer_helpers.R
\name{create_farm_overlap}
\alias{create_farm_overlap}
\alias{create_buffer_overlap_map}
\title{Title}
\usage{
create_farm_overlap(grid, farms)

create_buffer_overlap_map(buffer_overlap)
}
\arguments{
\item{buffer_overlap}{}
}
\description{
Title

Title
}
\examples{
Define source- and target-buffer indices and weights
}
