% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer_helpers.R
\name{create_buffer_overlap}
\alias{create_buffer_overlap}
\title{Title}
\usage{
create_buffer_overlap(grid, buffer)
}
\arguments{
\item{buffer}{}
}
\description{
Title
}
\examples{
Define source- and target-buffer indices and weights
}
