% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_triangle_grid.R
\name{create_custom_triangle_grid}
\alias{create_custom_triangle_grid}
\title{Create triangle grid}
\usage{
create_custom_triangle_grid(
  landscape,
  cellarea = NULL,
  n = NULL,
  middle = FALSE,
  celltype = c("triangle"),
  offset = c("corner", "middle", "bottom", "left"),
  orientation = c("diag", "anti")
)
}
\description{
Due to \code{sf::st_triangulate} / \code{sf::st_triangulate_constrained} producing
differently oriented triangles on a unstructured square grid, there is a
need for an approach, that returns consisting triangles throughout.
}
\note{
This function is meant to be inner function for \code{\link{create_grid}}, therefore
its parameters are broad.
}
