% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{create_grid}
\alias{create_grid}
\title{Title}
\usage{
create_grid(
  landscape,
  cellarea = NULL,
  n = NULL,
  middle = FALSE,
  center_as_centroid = FALSE,
  celltype = c("square", "hexagon", "hexagon_rot", "triangle", "triangulate_constrained"),
  offset = c("corner", "middle", "bottom", "left")
)
}
\arguments{
\item{landscape}{Landscape as \code{sf} object}

\item{cellarea}{Positive number that is less than or equal to \code{st_area(landscape)}.}

\item{n}{Integer specifying how many cells to place horizontally and vertically.}

\item{middle}{Offset the grid so equal sized cells are in the middle.}

\item{center_as_centroid}{Ensure that there is a cell, whose centroid is the center of \code{landscape}.}

\item{celltype}{Cell shapes \code{c("square", "hexagon", "hexagon_rot", "triangle")}}

\item{offset}{Aligned with initial cell placed at lower left corner}
}
\description{
Title
}
\note{
In \code{offset}, \code{"bottom"} means aligned to bottom;
Thus bottom middle.

⬛⬛⬛

⚫⚫⬛

⚫⚫⬛
}
