% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_si_model.R
\name{create_si_model}
\alias{create_si_model}
\title{Deterministic SI model with pair-wise infectious interactions}
\usage{
create_si_model(grid, beta_matrix, init, overlap, root = c("B", "C", "none"))
}
\arguments{
\item{grid}{Represents a grid of a landscape}

\item{beta_mat}{Infection rate matrix}

\item{y_init}{c(S, I) compartments for t = 0}

\item{target_overlap}{}
}
\description{
Deterministic SI model with pair-wise infectious interactions
}
\examples{
this follows the population from 0 until
tau (time when prevalence(target) = 50\%)
}
