% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{create_triangle_grid}
\alias{create_triangle_grid}
\title{Title}
\usage{
create_triangle_grid(n, cellsize, landscape)
}
\arguments{
\item{landscape}{}
}
\description{
Title
}
\note{
Will attempt to converge on \code{n} first, then \code{cellsize}
}
