% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/half_normal_kernel.R
\name{half_normal_kernel}
\alias{half_normal_kernel}
\title{Density for the half-normal distribution}
\usage{
half_normal_kernel(x, mean = 1, location = 0)
}
\arguments{
\item{x}{Numeric vector}

\item{mean}{mean of the half-normal distribution}
}
\description{
Density for the half-normal distribution
}
