% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_as_null.R
\name{na_as_null}
\alias{na_as_null}
\alias{null_as_na}
\title{Coerce \code{NA} into \code{NULL}}
\usage{
na_as_null(x)

null_as_na(x)
}
\arguments{
\item{x}{}
}
\description{
Note that \link[rlang:op-null-default]{rlang::\%||\%} treats \code{NA} as a value, i.e. \code{NA \%||\% 42 == NA}.
}
