% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_helpers.R
\name{create_farm_placement}
\alias{create_farm_placement}
\alias{get_buffer_source_target}
\alias{get_middle_buffer}
\title{Title}
\usage{
create_farm_placement(
  landscape,
  buffer_radius = 0.15,
  buffer_offset_percent = 0.2
)

get_buffer_source_target(
  landscape_width,
  landscape_height,
  buffer_radius = 5,
  buffer_offset_percent = 0.1
)

get_middle_buffer(source_target, buffer_radius)
}
\arguments{
\item{landscape}{an object created by \code{create_landscape}}

\item{buffer_offset_percent}{}
}
\description{
Title
}
\examples{
landscape <- create_landscape()
create_farm_placement(landscape,
                         buffer_radius = 5,
                         buffer_offset_percent = 0.1)

}
