% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_cellarea.R
\name{seq_cellarea}
\alias{seq_cellarea}
\title{Generate \code{cellarea} sequence}
\usage{
seq_cellarea(precision, min_cellarea, max_cellarea)
}
\arguments{
\item{precision}{Numeric between \code{0} and \code{1}}

\item{min_cellarea, max_cellarea}{}
}
\value{
Sequence between \code{min_cellarea} and \code{max_cellarea}
}
\description{
Geometric sequence of \code{cellarea} used in \link{create_grid} based
on increasing precision
}
