% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_blank_background.R
\name{theme_blank_background}
\alias{theme_blank_background}
\title{Removes background from \code{\link{ggplot}}-plots.}
\usage{
theme_blank_background(...)
}
\arguments{
\item{...Does}{nothing}
}
\value{
Theme
}
\description{
Removes background from \code{\link{ggplot}}-plots.
}
