% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{create_grid}
\alias{create_grid}
\title{Title}
\usage{
create_grid(
  landscape,
  cellarea = NULL,
  n = NULL,
  celltype = c("square", "hexagon", "hexagon_rot", "triangle"),
  offset = c("corner", "middle", "bottom", "left")
)
}
\arguments{
\item{landscape}{Landscape as \code{sf} object}

\item{cellarea}{Positive number that is less than or equal to \code{st_area(landscape)}.}

\item{celltype}{Cell shapes \code{c("square", "hexagon", "hexagon_rot", "triangle")}}

\item{offset}{Aligned with initial cell placed at lower left corner}
}
\description{
Title
}
\note{
In \code{offset}, \code{"bottom"} means aligned to bottom;
Thus bottom middle.

⬛⬛⬛

⚫⚫⬛

⚫⚫⬛
}
