% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{create_grid}
\alias{create_grid}
\title{Title}
\usage{
create_grid(n, landscape, landscape_scale)
}
\arguments{
\item{n}{Number of splits in}

\item{landscape}{Landscape as \code{sf} object}

\item{landscape_scale}{Width / height of the landscape}
}
\description{
Title
}
