% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grid.R
\name{create_grid}
\alias{create_grid}
\title{Title}
\usage{
create_grid(
  n,
  landscape,
  landscape_scale,
  offset = c("corner", "middle", "bottom", "left"),
  square = TRUE
)
}
\arguments{
\item{n}{Number of splits in}

\item{landscape}{Landscape as \code{sf} object}

\item{landscape_scale}{Width / height of the landscape}

\item{offset}{Aligned with initial cell placed at lower left corner}

\item{square}{}
}
\description{
Title
}
\note{
In \code{offset}, \code{"bottom"} means aligned to bottom;
Thus bottom middle.

⬛⬛⬛
⚫⚫⬛
⚫⚫⬛
}
