% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_transmission_matrix.R
\name{create_transmission_matrix}
\alias{create_transmission_matrix}
\title{Title}
\usage{
create_transmission_matrix(grid, kernel, ...)
}
\arguments{
\item{...}{parameter(s) passed to the kernel, after distance}
}
\description{
Title
}
