% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenario_helpers.R
\name{create_farm_placement}
\alias{create_farm_placement}
\alias{get_buffer_source_target}
\alias{get_middle_buffer}
\title{Title}
\usage{
create_farm_placement(landscape)

get_buffer_source_target(
  landscape_width,
  landscape_height,
  buffer_radius = 5,
  buffer_offset_percent = 0.1
)

get_middle_buffer(source_target, buffer_radius)
}
\arguments{
\item{buffer_offset_percent}{}
}
\description{
Title
}
\examples{
world_scale <- 17
get_buffer_source_target(landscape_width = world_scale,
                         landscape_height = world_scale,
                         buffer_radius = 5,
                         buffer_offset_percent = 0.1)

}
