% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_cellarea.R
\name{seq_cellarea}
\alias{seq_cellarea}
\title{Generate \code{cellarea} sequence}
\usage{
seq_cellarea(precision = NULL, n = NULL, min_cellarea, max_cellarea)
}
\arguments{
\item{n, precision}{provide either \code{n} or \code{precision}, where \code{n} entails number of cells between
\code{min_cellarea} and \code{max_cellarea}, and \code{precision} is \code{(1 - precision)^n} sequence}

\item{min_cellarea, max_cellarea}{Numeric}
}
\value{
Sequence of length \code{n + 1} between \code{min_cellarea} and \code{max_cellarea}
}
\description{
Geometric sequence of \code{cellarea} used in \link{create_grid} based
on increasing precision
}
